+++ zlt-business/backend-server/src/main/java/com/central/backend/controller/KpnTagController.java
@@ -2,6 +2,7 @@ package com.central.backend.controller;
 
 import java.util.Map;
 
+import cn.hutool.core.util.ObjectUtil;
 import com.central.backend.service.IKpnTagService;
 import com.central.common.annotation.LoginUser;
 import com.central.common.model.KpnTag;
@@ -48,6 +49,15 @@ public class KpnTagController {
     })
     @GetMapping
     public Result<PageResult> list(@RequestParam Map<String, Object> params,@LoginUser SysUser user) {
+        if (ObjectUtil.isEmpty(params)) {
+            return Result.failed("请求参数不能为空");
+        }
+        if (ObjectUtil.isEmpty(params.get("page"))) {
+            return Result.failed("分页起始位置不能为空");
+        }
+        if (ObjectUtil.isEmpty(params.get("limit"))) {
+            return Result.failed("分页结束位置不能为空");
+        }
         return Result.succeed(kpnTagService.findList(params,user));
     }
 
@@ -57,6 +67,9 @@ public class KpnTagController {
     @ApiOperation(value = "查询")
     @GetMapping("/{id}")
     public Result findUserById(@PathVariable Long id) {
+        if (ObjectUtil.isEmpty(id)) {
+            return Result.failed("ID不能为空");
+        }
         KpnTag model = kpnTagService.getById(id);
         return Result.succeed(model, "查询成功");
     }
@@ -67,6 +80,24 @@ public class KpnTagController {
     @ApiOperation(value = "新增or更新")
     @PostMapping
     public Result saveOrUpdateKpnTag(@RequestBody KpnTag kpnTag, @LoginUser SysUser user) {
+        if (ObjectUtil.isEmpty(kpnTag)) {
+            return Result.failed("请求参数不能为空");
+        }
+        if (ObjectUtil.isEmpty(kpnTag.getNameZh())) {
+            return Result.failed("中文名称不能为空");
+        }
+        if (ObjectUtil.isEmpty(kpnTag.getNameKh())) {
+            return Result.failed("柬文名称不能为空");
+        }
+        if (ObjectUtil.isEmpty(kpnTag.getNameEn())) {
+            return Result.failed("英文名称不能为空");
+        }
+        if (ObjectUtil.isEmpty(kpnTag.getCategoryName())) {
+            return Result.failed("所属分类不能为空");
+        }
+        if (ObjectUtil.isEmpty(kpnTag.getCategoryId())) {
+            return Result.failed("所属分类ID不能为空");
+        }
         return kpnTagService.saveOrUpdateKpnTag(kpnTag,user);
     }
 
@@ -76,6 +107,9 @@ public class KpnTagController {
     @ApiOperation(value = "删除")
     @DeleteMapping("/{id}")
     public Result delete(@PathVariable Long id) {
+        if (ObjectUtil.isEmpty(id)) {
+            return Result.failed("ID不能为空");
+        }
         return kpnTagService.removeKpnTag(id);
     }
 }
